+++ src/main/java/com/sdl/selenium/extjs6/grid/Scrollable.java
@@ -4,7 +4,7 @@ import com.google.common.base.Strings;
 import com.sdl.selenium.WebLocatorUtils;
 import com.sdl.selenium.web.IWebLocator;
 import com.sdl.selenium.web.WebLocator;
-import com.sdl.selenium.web.utils.RetryUtils;
+import com.sdl.selenium.web.utils.Retry;
 
 import java.time.Duration;
 
@@ -115,7 +115,7 @@ public interface Scrollable extends IWebLocator {
 
     default boolean scrollPageDownTo(WebLocator el) {
         el.setVisibility(true);
-        return RetryUtils.retry(50, () -> {
+        return Retry.retry(50, () -> {
             boolean isPresent = el.waitToRender(Duration.ofMillis(100), false);
             if (!isPresent && !isScrollBottom()) {
                 scrollPageDown();