+++ src/main/java/com/sdl/selenium/extjs6/notification/Notification.java
@@ -2,7 +2,7 @@ package com.sdl.selenium.extjs6.notification;
 
 import com.sdl.selenium.web.SearchType;
 import com.sdl.selenium.web.WebLocator;
-import com.sdl.selenium.web.utils.RetryUtils;
+import com.sdl.selenium.web.utils.Retry;
 import com.sdl.selenium.web.utils.Utils;
 
 import java.time.Duration;
@@ -36,7 +36,7 @@ public class Notification extends WebLocator {
     public boolean doClose() {
         WebLocator close = new WebLocator(this).setClasses("x-tool-close");
         Utils.sleep(350);
-        return RetryUtils.retry(3, () -> {
+        return Retry.retry(3, () -> {
             this.waitToRender(Duration.ofMillis(300), false);
             return close.doClick() || !this.waitToRender(Duration.ofMillis(100), false);
         });