+++ src/test/java/com/sdl/selenium/extjs6/button/RetryIntegrationTest.java
@@ -6,7 +6,7 @@ import com.sdl.selenium.extjs6.form.TextField;
 import com.sdl.selenium.extjs6.panel.Panel;
 import com.sdl.selenium.web.utils.Call;
 import com.sdl.selenium.web.utils.Result;
-import com.sdl.selenium.web.utils.RetryUtils;
+import com.sdl.selenium.web.utils.Retry;
 import com.sdl.selenium.web.utils.Utils;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
@@ -17,7 +17,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.core.Is.is;
 
-public class RetryUtilsIntegrationTest extends TestBase {
+public class RetryIntegrationTest extends TestBase {
 
     private final Panel loginPanel = new Panel(null, "Login");
     private final Button register = new Button(loginPanel, "Register");
@@ -34,14 +34,14 @@ public class RetryUtilsIntegrationTest extends TestBase {
 
     @Test
     void retryUntilOneIs() {
-        Result<Boolean> notPresent = RetryUtils.retryUntilOneIs(2, new Button(loginPanel, "Registerx")::isPresent, new Button(loginPanel, "Loginx")::isPresent);
+        Result<Boolean> notPresent = Retry.retryUntilOneIs(2, new Button(loginPanel, "Registerx")::isPresent, new Button(loginPanel, "Loginx")::isPresent);
         assertThat(notPresent.result(), is(false));
         assertThat(notPresent.position(), is(0));
     }
 
     @Test
     void retryUntilOneIsDetails() {
-        Result<Boolean> notPresent = RetryUtils.retryUntilOneIsDetails(2,
+        Result<Boolean> notPresent = Retry.retryUntilOneIsDetails(2,
                 new Call<>("Registerx", () -> new Button(loginPanel, "Registerx").isPresent()),
                 new Call<>("Loginx", () -> new Button(loginPanel, "Loginx").isPresent())
         );
@@ -51,7 +51,7 @@ public class RetryUtilsIntegrationTest extends TestBase {
 
     @Test
     void retryUntilOneIs1() {
-        Result<Boolean> present = RetryUtils.retryUntilOneIs(2,
+        Result<Boolean> present = Retry.retryUntilOneIs(2,
                 new Button(loginPanel, "Registerx")::isPresent,
                 new Button(loginPanel, "Login")::isPresent);
         assertThat(present.result(), is(true));
@@ -60,7 +60,7 @@ public class RetryUtilsIntegrationTest extends TestBase {
 
     @Test
     void retryUntilOneIsDetails1() {
-        Result<Boolean> present = RetryUtils.retryUntilOneIsDetails(2,
+        Result<Boolean> present = Retry.retryUntilOneIsDetails(2,
                 new Call<>("Registerx", new Button(loginPanel, "Registerx")::isPresent),
                 new Call<>("Login", new Button(loginPanel, "Login")::isPresent));
         assertThat(present.result(), is(true));
@@ -69,7 +69,7 @@ public class RetryUtilsIntegrationTest extends TestBase {
 
     @Test
     void retryUntilOneIs2() {
-        Result<Boolean> present2 = RetryUtils.retryUntilOneIs(2,
+        Result<Boolean> present2 = Retry.retryUntilOneIs(2,
                 new Button(loginPanel, "Registerx")::isPresent,
                 new Button(loginPanel, "Loginy")::isPresent,
                 new Button(loginPanel, "Login")::isPresent);
@@ -79,7 +79,7 @@ public class RetryUtilsIntegrationTest extends TestBase {
 
     @Test
     void retryUntilOneIsDetails2() {
-        Result<Boolean> present2 = RetryUtils.retryUntilOneIsDetails(2,
+        Result<Boolean> present2 = Retry.retryUntilOneIsDetails(2,
                 new Call<>("Registerx", new Button(loginPanel, "Registerx")::isPresent),
                 new Call<>("Loginy", new Button(loginPanel, "Loginy")::isPresent),
                 new Call<>("Login", new Button(loginPanel, "Login")::isPresent));
@@ -90,7 +90,7 @@ public class RetryUtilsIntegrationTest extends TestBase {
     @Test
     void retryUntilOneIs3() {
         userId.setValue("test");
-        Result<String> notPresent = RetryUtils.retryUntilOneIs(2, userId::getValue, password::getValue);
+        Result<String> notPresent = Retry.retryUntilOneIs(2, userId::getValue, password::getValue);
         assertThat(notPresent.result(), equalTo("test"));
         assertThat(notPresent.position(), is(1));
     }
@@ -98,7 +98,7 @@ public class RetryUtilsIntegrationTest extends TestBase {
     @Test
     void retryUntilOneIsDetails3() {
         userId.setValue("test");
-        Result<String> notPresent = RetryUtils.retryUntilOneIsDetails(2, new Call<>("userId", userId::getValue), new Call<>("password", password::getValue));
+        Result<String> notPresent = Retry.retryUntilOneIsDetails(2, new Call<>("userId", userId::getValue), new Call<>("password", password::getValue));
         assertThat(notPresent.result(), equalTo("test"));
         assertThat(notPresent.position(), is(1));
     }
@@ -107,7 +107,7 @@ public class RetryUtilsIntegrationTest extends TestBase {
     void retryUntilOneIs4() {
         userId.setValue("");
         password.setValue("test");
-        Result<String> present = RetryUtils.retryUntilOneIs(2, userId::getValue, password::getValue);
+        Result<String> present = Retry.retryUntilOneIs(2, userId::getValue, password::getValue);
         assertThat(present.result(), equalTo("test"));
         assertThat(present.position(), is(2));
     }
@@ -116,7 +116,7 @@ public class RetryUtilsIntegrationTest extends TestBase {
     void retryUntilOneIs5() {
         userId.setValue("");
         password.setValue("");
-        Result<String> present = RetryUtils.retryUntilOneIs(2, userId::getValue, password::getValue);
+        Result<String> present = Retry.retryUntilOneIs(2, userId::getValue, password::getValue);
         assertThat(present.result(), equalTo(""));
         assertThat(present.position(), is(0));
         assertThat(present.timeOut(), is(true));
@@ -126,7 +126,7 @@ public class RetryUtilsIntegrationTest extends TestBase {
     void retryUntilOneIs6() {
         userId.setValue("");
         password.setValue("");
-        Result<String> present = RetryUtils.retryUntilOneIs(Duration.ofSeconds(5), userId::getValue, password::getValue);
+        Result<String> present = Retry.retryUntilOneIs(Duration.ofSeconds(5), userId::getValue, password::getValue);
         assertThat(present.result(), equalTo(""));
         assertThat(present.position(), is(0));
         assertThat(present.timeOut(), is(true));
@@ -135,7 +135,7 @@ public class RetryUtilsIntegrationTest extends TestBase {
     @Test
     void retryUntilOneIs7() {
         userId.setValue("");
-        String present = RetryUtils.retry(Duration.ofSeconds(10), userId::getValue);
+        String present = Retry.retry(Duration.ofSeconds(10), userId::getValue);
         assertThat(present, equalTo(""));
     }
 }